% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_meansures.R
\name{get_features}
\alias{get_features}
\title{Calculate some summary measures}
\usage{
get_features(x, w, act_cols, h_sed = 100, h_mvpa = 2020)
}
\arguments{
\item{x}{placeholder}

\item{w}{placeholder}

\item{h_sed}{placeholder}
}
\description{
This function calculates fragmentation measures
}
\examples{
\dontrun{

}

}
\references{

}
