% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_meansures.R
\name{get_frag}
\alias{get_frag}
\title{Calculate fragmentation measures}
\usage{
get_frag(x, w, h_sed = 100)
}
\arguments{
\item{x}{placeholder}

\item{w}{placeholder}

\item{h_sed}{placeholder}
}
\description{
This function calculates fragmentation measures
}
\examples{
\dontrun{

}

}
\references{

}
