% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_meansures.R
\name{get_states}
\alias{get_states}
\title{Calculate Total weartime plus sedentary/mvpa time}
\usage{
get_states(x, w, h_sed = 100, h_mvpa = 2020)
}
\arguments{
\item{x}{placeholder}

\item{w}{placeholder}

\item{h_sed}{placeholder}

\item{h_mvpa}{placeholder}
}
\description{
This function calculates total activity volume
}
\examples{
\dontrun{

}

}
\references{

}
