% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_meansures.R
\name{get_TAC}
\alias{get_TAC}
\title{Calculate TAC/LTAC}
\usage{
get_TAC(x, w)
}
\arguments{
\item{x}{placeholder}

\item{w}{placeholder}
}
\description{
This function calculates total activity volume
}
\examples{
\dontrun{

}

}
\references{

}
