% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mortality_2011_D}
\alias{Mortality_2011_D}
\title{A dataset containing processed publicly linked mortality data for NHANES 2005-2006 participants. This data corresponds to the 2011 release of the linked mortality data.
As new mortality data is released this data file will be updated.}
\format{A data frame with 10,348 rows and 14 variables

\describe{
   \item{SEQN}{Unique subject identifier}
   \item{eligstat}{Eligibility status for mortality follow-up
         \describe{
                 \item{1}{ Eligible}
                 \item{2}{ Under age 18, not available for public release}
                 \item{3}{ Ineligible}
         }
   }
   \item{mortat}{Indicator for whether participant was found to be alive or deceased at follow-up time given by
   permth_exm and permth_int
         \describe{
                 \item{0}{ Assumed alive}
                 \item{1}{ Assumed deceased}
                 \item{NA}{ Under age 18, not available for public release or ineligible for mortality follow-up}
         }
   }
   \item{permth_exm}{Time in months from the mobile examination center (MEC) assessment where mortality was assessed.}
   \item{permth_int}{Time in months from the household interview where mortality was assessed.}
   \item{ucod_leading}{Underlying cause of death recode from UCOD_113 leading causes where available. Specific causes:
           \describe{
                 \item{001}{ Diseases of the heart (I00-I09, I11, I13, I20-I51)}
                 \item{002}{ Malignant neoplasms (C00-C97)}
                 \item{003}{ Chronic lower respiratory diseases (J40-J47)}
                 \item{004}{ Accidents (unintentional injuries) (V01-X59, Y85-Y86)}
                 \item{005}{ Cerebrovascular diseases (I60-I69)}
                 \item{006}{ Alzheimer's disease (G30)}
                 \item{007}{ Diabetes mellitus (E10-E14)}
                 \item{008}{ Influenza and pneumonia (J09-J18)}
                 \item{009}{ Nephritis, nephrotic syndrome and nephrosis (N00-N07, N17-N19, N25-N27)}
                 \item{010}{ All other causes (residual)}
                 \item{NA}{ Ineligible, under age 18, assumed alive or no cause data}
           }
     }
     \item{diabetes_mcod}{diabetes flag from multiple cause of death (mcod)}
     \item{hyperten_mcod}{hyperten flag from multiple cause of death (mcod)}
     \item{mortscrce_ndi}{mortality source: NDI match}
     \item{mortscrce_ssa}{mortality source: SSA information}
     \item{mortscrce_cms}{mortality source: CMS information}
     \item{mortscrce_dc}{mortality source: death certificate match}
     \item{mortscrce_dcl}{mortality source: data collection}

}}
\source{
\url{https://www.cdc.gov/nchs/data-linkage/mortality-public.htm}
}
\usage{
Mortality_2011_D
}
\description{
A dataset containing processed publicly linked mortality data for NHANES 2005-2006 participants. This data corresponds to the 2011 release of the linked mortality data.
As new mortality data is released this data file will be updated.
}
\keyword{datasets}
