% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mortality_D}
\alias{Mortality_D}
\title{A dataset containing processed publicly linked mortality data for NHANES 2005-2006 participants. This data corresponds to the 2011 release of the linked mortality data.
As new mortality data is released this data file will be updated.}
\format{A data frame with - rows and - variables
\describe{
   \item{SEQN}{Unique subject identifier}
}}
\source{
\url{https://www.cdc.gov/nchs/data-linkage/mortality-public.htm}
}
\usage{
Mortality_D
}
\description{
A dataset containing processed publicly linked mortality data for NHANES 2005-2006 participants. This data corresponds to the 2011 release of the linked mortality data.
As new mortality data is released this data file will be updated.
}
\keyword{datasets}
