% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesdata_package.R
\docType{package}
\name{nhanesdata-package}
\alias{nhanesdata-package}
\alias{nhanes}
\title{NHANES data package}
\description{
This package contains....
}
\section{Notation}{


Necessary?
}

\references{

}
