% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{process_accel}
\alias{process_accel}
\title{Process NHANES 2003-2004 and 2005-2006 accelerometry data}
\usage{
process_accel(names_accel_xpt = c("PAXRAW_C", "PAXRAW_D"),
  write = FALSE, local = FALSE, localpath = NULL, deleteraw = TRUE,
  urls = NULL)
}
\arguments{
\item{names_accel_xpt}{A character vector of names for the zipped raw files. These should be of the form PAXRAW_\* where \* corresponds to the
letter of the alphabet indicating which "wave" the data is from. For example, PAXRAW_C and PAXRAW_D correspond to the 2003-2004 and 2005-2006 waves
accelerometery data, respectively. A vector containing PAXRAW_C and PAXRAW_D is the default and will process both 2003-2004 and 2005-2006 waves data.}

\item{write}{logical argument indicating whether a .rda file should be created for each wave of processed data.
Defaults to FALSE. If TRUE, each wave of data will be saved in the location specified by the localpath variable. Each
wave will be saved as an .rda file named PAXINTEN_\*.rda where \* corresponds to the letter asspciated with a particular NHANES wave.}

\item{local}{logical argument indicating whether the zippped raw .xpt accelerometry files are stored locally.
If FALSE, will download the data into a temporary file from the CDC website and process the data. If TRUE,
the zipped data will be sourced locally. Defaults to FALSE.}

\item{localpath}{character string indicating where the locally zipped raw .xpt files are.}

\item{deleteraw}{logical argument indicating whether to delete the unzipped .xpt files after reading them into R.}
}
\description{
This function processes the raw NHANES 2003-2004 and 2005-2006 accelerometry data
as provided by the NHANES study on the CDC website.
Note that due to the large file size of the unzipped .xpt files, this function uses a non-trivial
amount of RAM (~12 GB peak). To avoid crashing your computer when running, please ensure you have
enough RAM available.
}
\examples{
\dontrun{
library("nhanesdata")
process_accel()
process_flags()
}

}
