% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{process_flags}
\alias{process_flags}
\title{Process wear/non-wear flags for NHANES 2003-2004 and 2005-2006 accelerometry data}
\usage{
process_flags(x, write = FALSE, localpath = NULL,
  days_distinct = FALSE, window = 90L, tol = 2L, tol_upper = 99L,
  nci = TRUE, ...)
}
\arguments{
\item{x}{A list with each element corresponding to a data matrix of activity counts in the 1440+ format with 7 rows per individual.
The data should be sorted by participant (SEQN) and then in descending order chronologically. The output of \code{\link{process_accel}} can be fed directly to
this argument. See examples.}

\item{write}{logical argument indicating whether a .rda file should be created for each wave of processed data.
Defaults to FALSE. If TRUE, each wave of data will be saved in the location specified by the localpath variable. Each
wave will be saved as an .rda file named Flags_\*.rda where \* corresponds to the letter asspciated with a particular NHANES wave.}

\item{localpath}{character string indicating where the locally processed .rda files should be saved if write = TRUE.}

\item{window}{size of the moving window used to assess non-wear in minutes. Defaults to 90 minutes.
See \code{\link{weartime}} for more details.}

\item{tol}{maximum number of minutes with counts greater than 0 within the window allowed before a particular
minute is considered "wear". That is, if for a given minute, the window around that minute has tol + 1
activity counts greater than 0, this miute is considered "wear".
See \code{\link{weartime}} for more details.}

\item{tol_upper}{maximium activity count for any minute within the window
Defaults to 99. That is, for a given minute, if the window contains any minutes with
activity counts greater than tol.upper, this minute is considered "wear".
See \code{\link{weartime}} for more details.}

\item{...}{aditional arguments to be passed to \code{\link{weartime}}.}
}
\value{
The function process_flags returns a list with number of elemnts equal to the number of elements in the object supplied to the "x" argument.
Each element of the list returned is a dataframe with the following columns:
}
\description{
This function creates wear/non-wear flag matrices for processed NHANES 2003-2004 and 2005-2006 accelerometry data
}
\examples{
\dontrun{
process_accel(write=FALSE)
}

}
\references{

}
