% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{process_mort}
\alias{process_mort}
\title{Process mortality data for NHANES 2003-2004 and 2005-2006 waves}
\usage{
process_mort(waves = c("C", "D"), mort_release_yr = 2011,
  write = FALSE, localpath = NULL)
}
\arguments{
\item{waves}{}

\item{mort_release_yr}{}

\item{write}{logical argument indicating whether a .rda file of wear/non-wear flags
should be created for each wave of processed data. Defaults to FALSE.}

\item{localpath}{}
}
\value{
This function returns a list
}
\description{
This function creates a clean mortality dataset which can be combined with data from the
NHANES 2003-2004/2005-2006 waves.
}
\examples{
\dontrun{


}

}
\references{
\url{https://www.cdc.gov/nchs/data-linkage/mortality-public.htm}
}
