% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnhanesdata_package.R
\docType{package}
\name{rnhanesdata-package}
\alias{rnhanesdata-package}
\alias{rnhanesdata}
\title{NHANES data package}
\description{
This package contains tools for processing the NHANES 2003-2004 and 2005-2006 waves' accelerometry data. Specifically, the code
contained herein will transform the "long format" accelerometry data which is available for download from the CDC, to the 1440+ format
with one row per subject-day, calculate wear/non-wear flags on the tranformed data, merge NHANES data from non-accelerometry sources, and
process national mortality data linked to NHANES participants. In addition to code for processing the 2003-2006 accerlerometry data,
this package contains code to: 1) process and merge NHANES data across multiple "raw" (.xpt) data files which are downloaded from the NHANES website; and
2) process linked mortality data for the NHANES 2003-2006 participants.
}
\details{
This package contains four primary data processing functions and two functions to aid in the analysis of accelerometry data:

\itemize{
    \item{Data processing functions}{
          \itemize{
                \item{\code{\link{process_accel}}}{}
                \item{\code{\link{process_flags}}}{}
                \item{\code{\link{process_covar}}}{}
                \item{\code{\link{process_mort}}}{}
           }
    }
    \item{Accelerometry analysis helper functions}{
          \itemize{
                \item{\code{\link{reweight_accel}}}{}
                \item{\code{\link{exclude_accel}}}{}
           }
    }
}

The processed data included in the package are broadly divided into four categories:
1) activity count data; 2) wear/non-wear flags; 3) additional NHANES data; 4) linked mortality data.

\itemize{
    \item{Activity count data}{
          \itemize{
                \item{\code{\link{PAXINTEN_C}}}{}
                \item{\code{\link{PAXINTEN_D}}}{}
          }
    }
    \item{Wear/non-wear flags}{
          \itemize{
                \item{\code{\link{Flags_C}}}{}
                \item{\code{\link{Flags_D}}}{}
          }
    }
    \item{Additional NHANES data}{
          \itemize{
                \item{\code{\link{Covariate_C}}}{}
                \item{\code{\link{Covariate_D}}}{}
          }
    }
    \item{Linked mortality data}{
          \itemize{
                \item{\code{\link{Mortality_C}}}{}
                \item{\code{\link{Mortality_D}}}{}
          }
    }
}

Finally, this package also contains raw data.
}
\references{

}
\author{
Andrew Leroux <aleroux2@jhu.edu>
Ciprian Crainiceanu
}
