% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesdata_package.R
\name{to be completed}
\alias{to be completed}
\title{Example dataset}
\format{to be complete
\describe{
  \item{SEQN}{placeholder}
}}
\description{
Simulated data from the CONTENT dataset. Data contains information on child growth
as measured by WHO defined Z-scores
as well as gender.
}
\references{

}
