% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{exclude_accel}
\alias{exclude_accel}
\title{Remove days with too few/too much weartime and NHANES data quality flags.}
\usage{
exclude_accel(act, flags, threshold_lower = 600, rm_PAXSTAT = TRUE,
  rm_PAXCAL = TRUE)
}
\arguments{
\item{act}{Activity data matrix}

\item{flags}{Wear/Non-wear flag matrix}

\item{rm_PAXSTAT}{Logical argument indicating whether to remove based on the data reliability indicator variable PAXSTAT. Defaults to TRUE.}

\item{rm_PAXCAL}{Logical argument indicating whether to remove based on the accelerometer calibration indicator variable PAXCAL. Defaults to TRUE.}

\item{threshold}{upper limit for number of non-wear minutes. Defaults to 600 minutes (10 hours).}

\item{return_act}{Not currently implemented.}
}
\description{
This function subsets accelerometry data by wear/non-wear time criteria, as well as
NHANES data quality flags.
}
\examples{
\dontrun{
process_accel(write=FALSE)
}

}
\references{

}
