% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accel.R
\name{process_covar}
\alias{process_covar}
\title{Merge non-accelerometry data for NHANES waves}
\usage{
process_covar(waves = c("C", "D"), varnames = c("SDDSRVYR", "WTMEC2YR",
  "WTINT2YR", "SDMVPSU", "SDMVSTRA", "RIDAGEMN", "RIDAGEEX", "RIDRETH1",
  "RIAGENDR", "BMXWT", "BMXHT", "BMXBMI", "DMDEDUC2", "ALQ101", "ALQ110",
  "ALQ120Q", "ALQ120U", "ALQ130", "SMQ020", "SMD030", "SMQ040", "MCQ220",
  "MCQ160F", "MCQ160B", "MCQ160C", "PFQ049", "PFQ054", "PFQ057", "PFQ059",
  "PFQ061B", "PFQ061C", "DIQ010"), localpath = NULL, write = FALSE,
  extractAll = FALSE)
}
\arguments{
\item{waves}{character vector with entries of (capitalized) letter of the alphabet corresponding to the
NHANES wave of interest. Defaults to a vector containing "C" and "D" corresponding to the NHANES 2003-2004 and 2005-2006 waves.}

\item{varnames}{character vector indicating which column names are to be searched for.
Will check all .XPT files in located in the directory specified by dataPath. If extractAll = TRUE, then this argument is effectively ignored. Defaults
to variables which are required to create the processed data matrices \code{\link{Covariate_C}} and  \code{\link{Covariate_D}}.}

\item{localpath}{file path where covariate data are saved. Covariate data must be in .XPT format,
and should be in their own folder. For example, PAXRAW_C.XPT should not be located in the folder with
your covariate files. This will not cause an error, but the code will take much longer to run.}

\item{write}{logical argument indicating whether a .rda file of covariate data. Defaults to FALSE.}

\item{extractAll}{logical argument indicating whether all columns of all .XPT files in the search path should be returned.
If extractALL = TRUE, all variables from all .XPT files with
Defaults to FALSE.}
}
\value{
This function will return a list with number of elements less than or equal to the number of waves of data specified by the "waves"
argument. The exact number of elements returned will depend on whether all files specified by the user are found in either: 1) the local directory
indicated by the localpath argument; or available in the data package. Each element of the list returned is a data frame. The columns of the data frame
will depend on the variables specified in the "varnames" argument and whether those variable are available for some (or any) of the waves specfied by the waves argument.
At a minimum, the returned dataframe will contain a column for each subject identifier (SEQN).

If the demographic datasets are not in the directory specified by localpath, there is no guarantee that all participants in a wave will be included in the retunred results.
}
\description{
This function retrieves and merges covariate data
from one or more NHANES data files across one or more waves of the study. Variables are merged using the NHANES unique subject identifier (SEQN).
}
\details{
This function will only recognize those files which have "SEQN" as their first column name as this is the format across all NHANES data files.
}
\examples{
\dontrun{


}

}
\references{

}
